Allow accessing properties in pluginRepositories block
This block was deliberately locked down in 2.14, not allowing access
to the `Settings` API during the initial pass over the settings script.
This was to prevent users from accessing methods that might not work as
expected in this early phase.

However, this has created a cognitive disconnect between how this block
works and how every other block in a .gradle file works. Most importantly,
users could not access any properties, which are often used to store credentials.

This change makes the block behave like every other configuration closure, i.e.
when a method or property is not found, it delegates up to the `SettingsScript` API.